refactor(adapter-server): split agui-resume-handler.test.ts + fix randomUUID import
- Split agui-resume-handler.test.ts (614 lines) into:
  - agui-resume-handler.test.ts (~360 lines): core execution path tests
    (happy path, cancelled, swapped action, digest mismatch, expiry, unsupported status)
  - agui-resume-handler-authz.test.ts (~260 lines): authz/replay tests
    (cross-user §6.2p5, unauthenticated §6.3, replay §6.2p4, approve-with-edits, permission slot §6.1/§6.4)
- Import randomUUID from node:crypto explicitly in agui-interrupt.ts (addresses
  Gemini review: consistent with createHash import, no reliance on global crypto)

Both new files remain under the 500-line project guideline.
